Update & move truncate functions to background
